fix: unknown ts extension 오류 해결
