+++ server/package.json
@@ -5,11 +5,10 @@
   "type": "module",
   "main": "index.js",
   "scripts": {
-    "start": "node dist/index.js",
+    "start": "ts-node src/index.ts",
     "build": "tsc -p tsconfig.json",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "server": "nodemon app.js",
-    "front": "cd ../client && npm start",
+    "server": "nodemon --watch 'src/**/*.ts' --exec ts-node ./src/index.ts",
     "dev": "concurrently \"npm run server\" \"npm run front\""
   },
   "author": "",
@@ -30,6 +29,7 @@
   "devDependencies": {
     "@types/express": "^4.17.17",
     "@types/node": "^18.14.6",
+    "ts-node": "^10.9.1",
     "typescript": "^4.9.5"
   }
 }
